# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FakeStore API Documentation",
        "description": "Swagger documentation for FakeStore API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/invoiceitem/create": {
            "post": {
                "tags": [
                    "Invoice Item"
                ],
                "summary": "Create a new invoice item",
                "description": "Creates a new invoice item record with invoice ID, product ID, quantity, and price.",
                "operationId": "createInvoiceItem",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "invoice_id",
                                    "product_id",
                                    "qty",
                                    "price"
                                ],
                                "properties": {
                                    "invoice_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "product_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "qty": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 3
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 150.5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Invoice item created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "new_data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "invoice_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "product_id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "qty": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 3
                                                },
                                                "price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 150.5
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2025-08-07T10:00:00.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2025-08-07T10:10:00.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "error": {
                                            "properties": {
                                                "invoice_id": {
                                                    "type": "string",
                                                    "example": "The invoice id field is required."
                                                },
                                                "product_id": {
                                                    "type": "string",
                                                    "example": "The product id field is required."
                                                },
                                                "qty": {
                                                    "type": "string",
                                                    "example": "The qty must be a number."
                                                },
                                                "price": {
                                                    "type": "string",
                                                    "example": "The price must be a number."
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "status_code": {
                                            "type": "integer",
                                            "example": 422
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/invoiceitem/update": {
            "post": {
                "tags": [
                    "Invoice Item"
                ],
                "summary": "Update an existing invoice item",
                "description": "Updates an invoice item record by given ID with new invoice ID, product ID, quantity, and price.",
                "operationId": "updateInvoiceItem",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id",
                                    "invoice_id",
                                    "product_id",
                                    "qty",
                                    "price"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "invoice_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "product_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "qty": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 5
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 250.75
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Invoice item updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "update_data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "invoice_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "product_id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "qty": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 5
                                                },
                                                "price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 250.75
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2025-08-07T10:00:00.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2025-08-07T12:00:00.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Invoice item not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "resource not found"
                                        },
                                        "status_code": {
                                            "type": "integer",
                                            "example": 404
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/invoiceitem/delete": {
            "post": {
                "tags": [
                    "Invoice Item"
                ],
                "summary": "Delete an invoice item by ID",
                "description": "Deletes an invoice item record by the given ID.",
                "operationId": "deleteInvoiceItem",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Invoice item deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "delete_data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "invoice_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "product_id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "qty": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 5
                                                },
                                                "price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 250.75
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2025-08-07T10:00:00.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2025-08-07T12:00:00.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Invoice item not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "resource not found"
                                        },
                                        "status_code": {
                                            "type": "integer",
                                            "example": 404
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}